# swagger.yaml
openapi: 3.0.0
info:
  title: Tourism
...
  contact: {}
  version: '1.0'
servers:
- url: https://tourisim.onrender.com/
  variables: {}
paths:
  /tours:
    get:
      tags:
      - Tours
      summary: Get ALL Tours
      description: Retrieve a list of all available tours.
      operationId: GetALLTours
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    post:
      tags:
      - Tours
      summary: Create New Tour
      description: Create a new tour with the specified details. Requires admin or lead-guide privileges.
      operationId: CreateNewTour
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateNewTourRequest'
              - example:
                  name: Tests Tour Average
                  duration: 1
                  maxGroupSize: 2
                  difficulty: easy
                  price: 397
                  summary: Test
                  imageCover: tour-1-cover.jpg
            example:
              name: Tests Tour Average
              duration: 1
              maxGroupSize: 2
              difficulty: easy
              price: 397
              summary: Test
              imageCover: tour-1-cover.jpg
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /users:
    get:
      tags:
      - Users
      summary: GET ALL USERS
      description: Retrieve a list of all users. Requires admin privileges.
      operationId: GETALLUSERS
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /reviews:
    get:
      tags:
      - Reviews
      summary: Get All Reviews
      description: Retrieve all reviews.
      operationId: GetAllReviews
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    post:
      tags:
      - Reviews
      summary: Create Reviews
      description: Create a new review for a tour.
      operationId: CreateReviews
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateReviewsRequest'
              - example:
                  review: Amazing!!
                  rating: 4
                  tour: 64e039e8771ee444f829988b
                  user: 5c8a1dfa2f8fb814b56fa181
            example:
              review: Amazing!!
              rating: 4
              tour: 64e039e8771ee444f829988b
              user: 5c8a1dfa2f8fb814b56fa181
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /bookings:
    get:
      tags:
      - Bookings
      summary: Get All Booking
      operationId: GetAllBooking
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    post:
      tags:
      - Bookings
      summary: Create New Booking
      operationId: CreateNewBooking
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
    CreateNewTourRequest:
      title: CreateNewTourRequest
      required:
      - name
      - duration
      - maxGroupSize
      - difficulty
      - price
      - summary
      - imageCover
      type: object
      properties:
        name:
          type: string
        duration:
          type: integer
          format: int32
        maxGroupSize:
          type: integer
          format: int32
        difficulty:
          type: string
        price:
          type: integer
          format: int32
        summary:
          type: string
        imageCover:
          type: string
      example:
        name: Tests Tour Average
        duration: 1
        maxGroupSize: 2
        difficulty: easy
        price: 397
        summary: Test
        imageCover: tour-1-cover.jpg
...
    CreateReviewsRequest:
      title: CreateReviewsRequest
      required:
      - review
      - rating
      - tour
      - user
      type: object
      properties:
        review:
          type: string
        rating:
          type: integer
          format: int32
        tour:
          type: string
        user:
          type: string
      example:
        review: Amazing!!
        rating: 4
        tour: 64e039e8771ee444f829988b
        user: 5c8a1dfa2f8fb814b56fa181
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...